+++ SimulationEngineLib/law/gravity/helper/GpuNewtonGravityHelper.cu
@@ -8,7 +8,7 @@
 
 __device__ 
 void radiusComponentKernelHelper(int idx, Particle** particles, Vector3D* accelerations, int betweenParticlesTriangularCount, double G, int vectorsProcessedTriangular) {
-	int x, y;
+	unsigned long long x, y;
 	MatrixMaths::getLowerTriangularCoordinates(idx + vectorsProcessedTriangular, &x, &y);
 	Vector3D devicePRadiusComponent = getRadiusComponent(particles[x], particles[y], G);
 	accelerations[idx] = -getAcceleration(particles[y]->mass, devicePRadiusComponent);
@@ -19,7 +19,7 @@ __device__
 void addAccelerationsKernelLowerHelper(int idx, Particle** particles, Vector3D* accelerations, int x0, int y, int n, int vectorsProcessedTriangular) {
 	int x = idx + x0;
 	if(x < n) { 
-		int radiusComponentIndex = MatrixMaths::getLowerTriangularIndex(x, y);
+		unsigned long long radiusComponentIndex = MatrixMaths::getLowerTriangularIndex(x, y);
 		runOnParticle(particles[x], accelerations[radiusComponentIndex - vectorsProcessedTriangular]);
 	} 
 }
@@ -28,7 +28,7 @@ __device__
 void addAccelerationsKernelUpperHelper(int idx, Particle** particles, Vector3D* accelerations, int x0, int y, int n, int vectorsProcessedTriangular, int particlesProcessed, int betweenParticlesTriangularCount) {
 	int x = idx + x0;
 	if(x < n && x >= particlesProcessed) { 
-		int radiusComponentIndex = MatrixMaths::getUpperTriangularIndex(x, y);
+		unsigned long long radiusComponentIndex = MatrixMaths::getUpperTriangularIndex(x, y);
 		runOnParticle(particles[x], accelerations[radiusComponentIndex - vectorsProcessedTriangular + betweenParticlesTriangularCount]);
 	} 
 }
